+++ tests/iterator/iterator_helpers.h
@@ -1,0 +1,8 @@
+
+extern void expect_iterator_items(
+	git_iterator *i,
+	int expected_flat,
+	const char **expected_flat_paths,
+	int expected_total,
+	const char **expected_total_paths);
+
